Casts. Corrected key regeneration.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.09@2560 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-misc.cc
@@ -803,10 +803,11 @@ void spoton_misc::retrieveSymmetricData
 
 		    neighborOid = query.value(1).toString();
 		    publicKey = query.value(2).toByteArray();
-		    symmetricKey.resize(symmetricKeyLength);
+		    symmetricKey.resize
+		      (static_cast<int> (symmetricKeyLength));
 		    symmetricKey = spoton_crypt::strongRandomBytes
 		      (symmetricKey.length());
-		    hashKey.resize(symmetricKeyLength);
+		    hashKey.resize(static_cast<int> (symmetricKeyLength));
 		    hashKey = spoton_crypt::strongRandomBytes
 		      (hashKey.length());
 		  }
